Actually works
Locally at least. The unicode characters are listed at the bottom of
the UI. We have not yet handled different users but we’re off to a good
start.

Weird problem though. I have to change the port every time I run the
server with this code. I don’t understand it.